tasks/Backend_2.2/blog/requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
Django==4.0.1
django-bootstrap4==21.2
django-crispy-forms==1.9.2
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2021.5
